# TreciasDarbas1 - Copy/TreciasDarbas1 - Copy/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8000","description":"Generated server url"}],"paths":{"/bands":{"get":{"tags":["band-controller"],"operationId":"all","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelBand"}}}}}},"post":{"tags":["band-controller"],"operationId":"createBand","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Band"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Band"}}}}}}},"/bands/{bandId}/members":{"get":{"tags":["member-controller"],"operationId":"allByBand","parameters":[{"name":"bandId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Member"}}}}}}},"post":{"tags":["member-controller"],"operationId":"createMemberForBand","parameters":[{"name":"bandId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Member"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Member"}}}}}}},"/bands/{bandId}":{"get":{"tags":["band-controller"],"operationId":"one","parameters":[{"name":"bandId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelBand"}}}}}}}},"components":{"schemas":{"Band":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"genre":{"type":"string"},"date":{"type":"string"}}},"Member":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"}}},"CollectionModelEntityModelBand":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"bandList":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelBand"}}}},"_links":{"$ref":"#/components/schemas/Links"}}},"EntityModelBand":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"genre":{"type":"string"},"date":{"type":"string"},"_links":{"$ref":"#/components/schemas/Links"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}